refactor(firestore): Refactor firestore/retrieve.go
Split firestore/retrieve.go samples to separate file, widen region tags
