+++ firestore/retrieve_data_get_as_map.go
@@ -1,0 +1,36 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package firestore
+
+// [START firestore_data_get_as_map]
+
+import (
+	"context"
+	"fmt"
+
+	"cloud.google.com/go/firestore"
+)
+
+func docAsMap(ctx context.Context, client *firestore.Client) (map[string]interface{}, error) {
+	dsnap, err := client.Collection("cities").Doc("SF").Get(ctx)
+	if err != nil {
+		return nil, err
+	}
+	m := dsnap.Data()
+	fmt.Printf("Document data: %#v\n", m)
+	return m, nil
+}
+
+// [END firestore_data_get_as_map]
