+++ firestore/retrieve_data_get_dataset.go
@@ -1,0 +1,45 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package firestore
+
+// [START firestore_data_get_dataset]
+
+import (
+	"context"
+
+	"cloud.google.com/go/firestore"
+)
+
+func prepareRetrieve(ctx context.Context, client *firestore.Client) error {
+	cities := []struct {
+		id string
+		c  City
+	}{
+		{id: "SF", c: City{Name: "San Francisco", State: "CA", Country: "USA", Capital: false, Population: 860000}},
+		{id: "LA", c: City{Name: "Los Angeles", State: "CA", Country: "USA", Capital: false, Population: 3900000}},
+		{id: "DC", c: City{Name: "Washington D.C.", Country: "USA", Capital: true, Population: 680000}},
+		{id: "TOK", c: City{Name: "Tokyo", Country: "Japan", Capital: true, Population: 9000000}},
+		{id: "BJ", c: City{Name: "Beijing", Country: "China", Capital: true, Population: 21500000}},
+	}
+	for _, c := range cities {
+		_, err := client.Collection("cities").Doc(c.id).Set(ctx, c.c)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// [END firestore_data_get_dataset]
